pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "inmydata MCP server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "inmydata-openedge>=0.0.1",
    "mcp>=1.15.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "fastmcp==2.12.4",
    "uvicorn"
]
